Reorganized code to make it easier to find things
`hf.rs` should contain all things related to getting the wireshark
header fields set up and accessable.

`wrapperdefs.rs` contains definitions which make it easier to work with
the wireshark interface
