+++ dissect-netwayste/src/lib.rs
@@ -45,11 +45,15 @@ use std::io::{Error, ErrorKind};
 use std::mem;
 use std::net::SocketAddr;
 use std::os::raw::{c_int, c_void};
-use std::ptr;
 use std::sync::Mutex;
 
 mod netwaysteparser;
+mod hf;
+mod wrapperdefs;
+
 use netwaysteparser::{parse_netwayste_format, Sizing, VariableContainer, NetwaysteDataFormat::{self, Enumerator, Structure}};
+use hf::*;
+use wrapperdefs::*;
 
 /// Wireshark C bindings
 mod ws {
@@ -58,16 +62,9 @@ mod ws {
 
 // TODO get this from somewhere else. Not sure if self definition is the best route?
 const UDP_MTU_SIZE: usize = 1460;
-// PR_GATE reevaluate this once other fields are added in
-const ENUM_SIZE: i32 = mem::size_of::<i32>() as i32;
-
-// :( https://stackoverflow.com/questions/33850189/how-to-publish-a-constant-string-in-the-rust-ffi
-#[repr(C)]
-pub struct StaticCString(*const u8);
-unsafe impl Sync for StaticCString {}
 
 #[no_mangle]
-pub static plugin_version: StaticCString = StaticCString(b"0.0.2\0" as *const u8);
+pub static plugin_version: StaticCString = StaticCString(b"0.0.3\0" as *const u8);
 pub static NONE_STRING: StaticCString = StaticCString(b"None\0" as *const u8);
 /// Wireshark major & minor version
 #[no_mangle]
@@ -81,122 +78,7 @@ static mut plug_conwayste: ws::proto_plugin = ws::proto_plugin {
 };
 
 static mut proto_conwayste: c_int = -1;
-
 static mut ett_conwayste: c_int = -1;
-
-/// HFFieldAllocator keeps track of which header fields have been used during header field registration.
-///
-/// # Notes
-/// Internally it uses a run-time populated vector sized to the number of Netwayste enums/structures
-/// and their member fields. Assignment involves tracking which fields each slot has been assigned to.
-///
-/// # Panics
-/// When all header fields have been used up. This may occur if the number of registrations exceed
-/// the number of enum/struct members found during parsing.
-///
-#[derive(Debug)]
-struct HFFieldAllocator {
-    hf_fields: Vec<c_int>,
-    allocated: HashMap<CString, usize>,
-    options_map: HashMap<CString, CString>,
-}
-
-impl HFFieldAllocator {
-    fn new() -> HFFieldAllocator {
-        HFFieldAllocator {
-            hf_fields: Vec::new(),
-            allocated: HashMap::new(),
-            options_map: HashMap::new(),
-        }
-    }
-
-    /// Retrieves a pointer to the (mutable) allocated header field for the provided string.
-    ///
-    /// # Panics
-    /// Will panic if the provided String is not registered. This is intentional as a means to catch
-    /// bugs.
-    fn get(&mut self, name: &CString) -> &mut c_int {
-        if let Some(index) =  self.allocated.get(name) {
-            assert!(*index < self.hf_fields.len());
-            // Unwrap safe b/c of assert
-            let item = self.hf_fields.get_mut(*index).unwrap();
-            return item;
-        }
-        unreachable!();
-    }
-
-    /// Registers the provided string with the allocator. This must be called prior to any `get()`
-    /// calls!
-    fn register(&mut self, name: CString) {
-        // Add in a value of -1. Wireshark will overwrite this later.
-        self.hf_fields.push(-1);
-        // Map the index into the `hf_fields` vector to the field name
-        self.allocated.insert(name, self.hf_fields.len() - 1);
-    }
-
-    /// Creates a new optional entry for fields that are `Option<T>`. This entry is used to indicate
-    /// the field is None in the Netwayste decomposition tree.
-    fn new_option(&mut self, name: &CString) {
-        let new_name = format!("{}.option", name.to_str().unwrap());
-        let new_name_cstr = CString::new(new_name).unwrap();
-        self.options_map.insert(name.clone(), new_name_cstr.clone());
-
-        // register the header field for this "blahblah.option" to create the ID
-        self.register(new_name_cstr);
-    }
-
-    /// Retrieves the `CString` used when the provided field name is a `None` variant of `Option`
-    fn get_option_name(&self, name: &CString) -> &CString {
-        self.options_map.get(name).unwrap()
-    }
-
-    /// Retrieves the header field ID linked the provided field name when it is a `None` variant of
-    /// `Option`.
-    fn get_option_id(&mut self, name: &CString) -> &mut c_int {
-        let optioned_name = self.options_map.get(name).unwrap().clone();
-        self.get(&optioned_name)
-    }
-}
-
-// same as hf_register_info from bindings.rs except the pointer is *const instead of *mut
-// UGLY HACK
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct sync_hf_register_info {
-    pub p_id: *mut c_int,
-    pub hfinfo: ws::header_field_info,
-}
-
-unsafe impl Sync for sync_hf_register_info {}
-unsafe impl Send for sync_hf_register_info {}
-
-impl Default for ws::header_field_info {
-    fn default() -> Self {
-        ws::header_field_info {
-            name:   ptr::null(), // < [FIELDNAME] full name of this field
-            abbrev: ptr::null(), // < [FIELDABBREV] abbreviated name of this field
-            type_: FieldType::NoType as u32, // < [FIELDTYPE] field type
-            display: FieldDisplay::NoDisplay as i32, // < [FIELDDISPLAY] Base representation on display
-            strings: ptr::null(),
-            bitmask: 0, // < [BITMASK] bitmask of interesting bits
-            blurb: ptr::null(),   // < [FIELDDESCR] Brief description of field
-            id: -1,   // < Field ID
-            parent: -1,   // < parent protocol tree
-            ref_type: 0,    // < is this field referenced by a filter
-            same_name_prev_id: -1,   // < ID of previous hfinfo with same abbrev
-            same_name_next: ptr::null_mut(), // < Link to next hfinfo with same abbrev
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone)]
-struct sync_named_packet_types {
-    index: c_int,
-    name: *const i8,
-}
-
-unsafe impl Sync for sync_named_packet_types {}
 
 struct ConwaysteProtocolStrings {
     proto_full_name: CString,
@@ -216,62 +98,6 @@ impl ConwaysteProtocolStrings {
     }
 }
 
-#[repr(u32)]
-enum FieldDisplay {
-    NoDisplay = ws::field_display_e_BASE_NONE,
-    Decimal = ws::field_display_e_BASE_DEC,
-    Hexadecimal = ws::field_display_e_BASE_HEX,
-    Oct = ws::field_display_e_BASE_OCT,
-    DecHex = ws::field_display_e_BASE_DEC_HEX,
-    HexDec = ws::field_display_e_BASE_HEX_DEC,
-    Str = ws::field_display_e_STR_UNICODE,
-}
-
-#[derive(Debug)]
-#[repr(u32)]
-enum FieldType {
-    NoType = ws::ftenum_FT_NONE,
-    Char = ws::ftenum_FT_CHAR,
-    U8 = ws::ftenum_FT_UINT8,
-    U16 = ws::ftenum_FT_UINT16,
-    U32 = ws::ftenum_FT_UINT32,
-    U64 = ws::ftenum_FT_UINT64,
-    I8 = ws::ftenum_FT_INT8,
-    I16 = ws::ftenum_FT_INT16,
-    I32 = ws::ftenum_FT_INT32,
-    I64 = ws::ftenum_FT_INT64,
-    F32 = ws::ftenum_FT_FLOAT,
-    F64 = ws::ftenum_FT_DOUBLE,
-    Str = ws::ftenum_FT_STRING,
-    Str_z = ws::ftenum_FT_STRINGZ,
-}
-
-impl From<String> for FieldType {
-    fn from(input: String) -> Self {
-        input.as_str().into()
-    }
-}
-
-impl From<&str> for FieldType {
-    fn from(input: &str) -> Self {
-        use FieldType::*;
-        match input {
-            "u64" => U64,
-            "i64" => I64,
-            "u32" => U32,
-            "i32" => I32,
-            "u16" => U16,
-            "i16" => I16,
-            "u8" => U8,
-            "i8" => I8,
-            "f64" => F64,
-            "f32" => F32,
-            "char" => Char,
-            "String" => Str,
-            _ => NoType,
-        }
-    }
-}
 
 lazy_static! {
     static ref protocol_strings: ConwaysteProtocolStrings = ConwaysteProtocolStrings::new();
@@ -317,48 +143,9 @@ lazy_static! {
     };
 
     // setup protocol field array
-    static ref hf: Mutex<Vec<sync_hf_register_info>> = Mutex::new(Vec::new());
-}
+    static ref hf_info: Mutex<Vec<sync_hf_register_info>> = Mutex::new(Vec::new());
 
-// *************************************************************************************************
-// The following private functions are intended to be the only means to work with `hf_fields` and `hf`
-// static variables. Due to the non-definable order of `lazy-static` instantiation, both fields are
-// initialized when Wireshark registers the dissector, but before any dissection occurs. These
-// static variables use a `Mutex` and the usage of these functions ensure the locks are dropped cleanly.
-// We aren't multithreading here but it's still good practice and helps with readability.
-// (Suggested by https://users.rust-lang.org/t/how-can-i-use-mutable-lazy-static/3751/5)
-
-fn hf_register(name: CString) {
-    hf_fields.lock().unwrap().register(name);
-}
-
-fn hf_get(name: &CString) -> *mut c_int {
-    hf_fields.lock().unwrap().get(name) as *mut c_int
-}
-
-fn hf_new_option(name: &CString) {
-    let mut lock = hf_fields.lock().unwrap();
-    lock.new_option(name);
-}
-
-fn hf_get_option(name: &CString) -> *const i8 {
-    let lock = hf_fields.lock().unwrap();
-    lock.get_option_name(name).as_ptr()
-}
-
-fn hf_get_option_id(name: &CString) -> *mut c_int {
-    let mut lock = hf_fields.lock().unwrap();
-    lock.get_option_id(name) as *mut c_int
-}
-
-fn hf_as_ptr() -> *const sync_hf_register_info {
-    let ptr = hf.lock().unwrap().as_ptr();
-    ptr
-}
-
-fn hf_len() -> usize {
-    let len = hf.lock().unwrap().len();
-    len
+    static ref handoff_match_name: CString = { CString::new("udp.port").unwrap() };
 }
 
 // Just a sad little utility function to print hex in a u8 slice
@@ -369,23 +156,6 @@ fn print_hex(buf: &[u8]) {
     }
     let s: String = v.join("");
     println!("[{}]", s);
-}
-
-#[repr(u32)]
-enum WSColumn {
-    Protocol = ws::COL_PROTOCOL,
-    Info = ws::COL_INFO,
-}
-
-
-#[repr(u32)]
-#[derive(Debug, Copy, Clone)]
-enum WSEncoding {
-    BigEndian = ws::ENC_BIG_ENDIAN,
-    LittleEndian = ws::ENC_LITTLE_ENDIAN,
-    UTF8 = ws::ENC_UTF_8,
-    UTF16 = ws::ENC_UTF_16,
-    UTF8String = ws::ENC_LITTLE_ENDIAN + ws::ENC_UTF_8,
 }
 
 struct ConwaysteTree {
@@ -522,44 +292,6 @@ impl ConwaysteTree {
     }
 }
 
-/// A safe wrapper for `col_add_str`, which copies the provided string to the target column.
-fn column_add_str(pinfo: *mut ws::packet_info, column: WSColumn, name: CString) {
-    unsafe { ws::col_add_str((*pinfo).cinfo, column as i32, name.as_ptr()); }
-}
-
-/// A safe wrapper for `col_set_str`, which takes a pointer to the provided string and therefore
-/// must live for the duration of usage!
-fn column_set_str(pinfo: *mut ws::packet_info, column: WSColumn, name: &CString) {
-    unsafe { ws::col_set_str((*pinfo).cinfo, column as i32, name.as_ptr()); }
-}
-
-/// A safe wrapper for `col_clear`, which clears the specified column.
-fn column_clear(pinfo: *mut ws::packet_info, column: WSColumn) {
-    unsafe { ws::col_clear((*pinfo).cinfo, column as i32); }
-}
-
-// For an explanation of the difference of captured vs reported tvb lengths,
-// see https://seclists.org/wireshark/2015/Sep/15
-// For our purposes, this should match the reported length of the tvb because we aren't snapshotting
-// in the buffer. We want the entire packet.
-/// The number of bytes captured from a packet in the tv buffer.
-fn tvb_captured_length(tvb: *mut ws::tvbuff_t) -> i32 {
-    unsafe {
-        let len = ws::tvb_captured_length(tvb) as i32;
-        assert!(len > 0);
-        len
-    }
-}
-
-/// The number of bytes in the the entire tv buffer.
-fn tvb_reported_length(tvb: *mut ws::tvbuff_t) -> usize {
-    unsafe {
-        let len = ws::tvb_reported_length(tvb) as i32;
-        assert!(len > 0); // a length of zero means we're at the end of the buffer
-        len as usize
-    }
-}
-
 /// Decode packet bytes from the tv buffer into a netwayste packet
 fn get_cwte_packet(tvb: *mut ws::tvbuff_t) -> Result<NetwaystePacket, std::io::Error> {
     let tvblen = tvb_reported_length(tvb) as usize;
@@ -609,7 +341,6 @@ extern "C" fn dissect_conwayste(
 
     let conwayste_tree = ConwaysteTree::new(tvb, tree);
     conwayste_tree.decode(tvb);
-        // TODO: auto-generate more trees and items from inspecting AST of Packet
 
     // return the entire packet lenth.
     let captured_len = tvb_captured_length(tvb);
@@ -630,8 +361,8 @@ extern "C" fn proto_register_conwayste() {
     println!("called proto_register_conwayste()");
 
     // PR_GATE: See if it makes sense to combine these two routines into one
-    register_header_fields();
-    build_header_field_array();
+    hf::register_header_fields();
+    hf::build_header_field_array();
 
     unsafe {
         proto_conwayste = ws::proto_register_protocol(
@@ -642,168 +373,11 @@ extern "C" fn proto_register_conwayste() {
 
         ws::proto_register_field_array(
             proto_conwayste,
-            hf_as_ptr() as *mut ws::hf_register_info,
-            hf_len() as i32,
+            hf_info_as_ptr() as *mut ws::hf_register_info,
+            hf_info_len() as i32,
         );
         ws::proto_register_subtree_array(ett.as_ptr() as *const *mut i32, ett.len() as i32);
     }
-}
-
-/// For every enum/structure found by parsing `netwayste/src/net.rs` must have a header field identifier
-/// that Wireshark uses to refer to it. This routine will walk through the parsed-and-gutted
-/// `net.rs` and assign a header field ID to each one. It does this via registration with the header
-/// field allocator.
-fn register_header_fields() {
-    // Reserve a header field for the variant
-    for key in netwayste_data.keys() {
-        hf_register(key.clone());
-    }
-
-    for datastruct in netwayste_data.values() {
-        // Reserve a header field for each variant's fields
-        match datastruct {
-            Enumerator(_enums, fields) => {
-                // Reserve a header field for its fields.
-                for vfield in fields.values() {
-                    for vf in vfield.iter() {
-                        hf_register(vf.name.clone());
-
-                        // Check if there's an Option, and register an additional HF for it
-                        for format in &vf.format {
-                            match format {
-                                Sizing::Variable(VariableContainer::Optional) => {
-                                    hf_new_option(&vf.name);
-                                }
-                                _ => {
-                                    // Not of any concern, keep looking
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            Structure(fields) => {
-                // Reserve a header field for structure's fields
-                for f in fields {
-                    // Stuctures are *always* named so unwrap is safe.
-                    hf_register(f.name.clone());
-                }
-            }
-        }
-    }
-}
-
-// Walks the parsed `net.rs` AST and builds a header field entry for each enum, variants with data,
-// and structures. The header field entry is provided to Wireshark so that it knows how to interpret
-// each data field when it's added to the ett during packet dissection.
-fn build_header_field_array() {
-    let mut _hf = {
-        let mut _hf = vec![];
-        for enum_name in netwayste_data.keys() {
-            let f = hf_get(enum_name);
-
-            let enum_hf = sync_hf_register_info {
-                p_id: f,
-                hfinfo: ws::header_field_info {
-                    name:       enum_name.as_ptr() as *const i8,
-                    abbrev:     enum_name.as_ptr() as *const i8,
-                    type_:      FieldType::U32 as u32,
-                    display:    FieldDisplay::Decimal as i32,
-                    strings:    if let Some(strings) = enum_strings.get(enum_name) {
-                                    strings.as_ptr() as *const c_void
-                                } else {
-                                    ptr::null()
-                                },
-                    ..Default::default()
-                },
-            };
-            _hf.push(enum_hf);
-        }
-
-        for datastruct in netwayste_data.values() {
-            match datastruct {
-                Enumerator(_enums, fields) => {
-                    for vfield in fields.values() {
-                        for vf in vfield.iter() {
-                            let mut field_data_type = FieldType::Str;
-                            let mut field_display: FieldDisplay = FieldDisplay::Str;
-                            for fmt in vf.format.iter() {
-                                match fmt {
-                                    Sizing::Structure(s) => {
-                                        // TODO:
-                                    },
-                                    Sizing::Variable(VariableContainer::Optional) => {
-                                        // nothing to do, will use nested type
-                                        field_display = FieldDisplay::Str;
-
-                                        let f = hf_get_option_id(&vf.name);
-                                        let optioned_name = hf_get_option(&vf.name);
-                                        let variant_hf = sync_hf_register_info {
-                                            p_id: f,
-                                            hfinfo: ws::header_field_info {
-                                                name:       optioned_name,
-                                                abbrev:     optioned_name,
-                                                type_:      FieldType::Str as u32,
-                                                display:    FieldDisplay::Str as i32,
-                                                ..Default::default()
-                                            },
-                                        };
-                                        _hf.push(variant_hf);
-                                    }
-                                    Sizing::Variable(VariableContainer::Vector) => {
-                                        // nothing to do, will default to string if no further nesting
-                                    }
-                                    Sizing::Fixed(bytes) => {
-                                        field_display = FieldDisplay::Decimal;
-                                        match bytes {
-                                            8 => field_data_type = FieldType::U64,
-                                            4 => field_data_type = FieldType::U32,
-                                            2 => field_data_type = FieldType::U16,
-                                            1 => field_data_type = FieldType::U8,
-                                            // We shouldn't get other values here
-                                            unknown_byte_count @ _ => {
-                                                println!("Unknown byte count observed during header
-                                                    field construction: {}", unknown_byte_count);
-                                                field_data_type = FieldType::U64;
-                                            },
-                                        }
-                                        break;
-                                    }
-                                }
-                            }
-
-                            let f = hf_get(&vf.name);
-                            let variant_hf = sync_hf_register_info {
-                                p_id: f,
-                                hfinfo: ws::header_field_info {
-                                    name:       vf.name.as_ptr() as *const i8,
-                                    abbrev:     vf.name.as_ptr() as *const i8,
-                                    type_:      field_data_type as u32,
-                                    display:    field_display as i32,
-                                    ..Default::default()
-                                },
-                            };
-                            _hf.push(variant_hf);
-                        }
-                    }
-                }
-                Structure(fields) => {
-                    for f in fields {
-                        // TODO:
-                    }
-                }
-            }
-        }
-
-        _hf
-    };
-
-    // Append is only performed once so I am not wrapping it into its own function like the others
-    hf.lock().unwrap().append(&mut _hf);
-}
-
-lazy_static! {
-    static ref handoff_match_name: CString = { CString::new("udp.port").unwrap() };
 }
 
 /// Notifies Wireshark to call the dissector when finding UDP traffic on `ws::CONWAYSTE_PORT`.
